refactor(io): split checkpoint format into io::checkpoint module
Extracts CheckpointRecorder, CheckpointArchive, and the
checkpoint-{epoch:06} scanning/serialization logic out of io::run
into a dedicated io::checkpoint module, keeping TrainingRun focused
on the run directory handle (create/open/recorder/trim_after/archive).
