+++ cli/src/actions.rs
@@ -3,8 +3,8 @@ use nrn::activations::RELU;
 use nrn::charts::RenderConfig;
 use nrn::data::Dataset;
 use nrn::data::scalers::ScalerMethod;
+use nrn::io::checkpoint::CheckpointArchive;
 use nrn::io::png::save_rgb;
-use nrn::io::run::CheckpointArchive;
 use nrn::io::scalers::ScalerRecord;
 use nrn::model::{NeuralNetwork, NeuronLayerSpec};
 use std::error::Error;
