+++ core/src/io/mod.rs
@@ -1,4 +1,5 @@
 pub mod bytes;
+pub mod checkpoint;
 pub mod classes;
 pub mod data;
 pub mod gif;
