+++ core/src/io/run.rs
@@ -1,23 +1,12 @@
-use crate::evaluation::{Evaluation, EvaluationSet};
-use crate::evaluation_history::{EpochEvaluation, EvaluationHistory};
-use crate::io::bytes::secure_read;
+use crate::io::checkpoint::{CheckpointArchive, CheckpointRecorder, scan_checkpoints};
 use crate::io::hyperparams::HyperParamsRecord;
 use crate::io::json;
-use crate::io::optimizer as optimizer_io;
 use crate::io::path::PathExt;
-use crate::io::tensors;
-use crate::model::NeuralNetwork;
-use crate::optimizers::{Optimizer, OptimizerState};
-use crate::schedulers::Scheduler;
-use crate::training::TrainingCallback;
-use safetensors::SafeTensors;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use std::fs;
-use std::io::ErrorKind::{AlreadyExists, InvalidData};
+use std::io::ErrorKind::AlreadyExists;
 use std::io::{Error, Result};
 use std::path::{Path, PathBuf};
-use std::result::Result as StdResult;
 
 /// Top-level metadata for a training run directory.
 /// Written once by [`TrainingRun::create`] into `meta.json`.
@@ -38,85 +27,6 @@ impl TrainingMeta {
         let dir = Path::combine_safe_with_cwd(dir)?;
         json::save(self, dir.join("meta"))
     }
-}
-
-#[derive(Serialize, Deserialize)]
-struct CheckpointEvaluationSet {
-    train: CheckpointEvaluation,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    validation: Option<CheckpointEvaluation>,
-    test: CheckpointEvaluation,
-}
-
-#[derive(Serialize, Deserialize)]
-struct CheckpointEvaluation {
-    loss: f32,
-    accuracy: f32,
-}
-
-impl From<Evaluation> for CheckpointEvaluation {
-    fn from(eval: Evaluation) -> Self {
-        CheckpointEvaluation {
-            loss: eval.loss,
-            accuracy: eval.accuracy,
-        }
-    }
-}
-
-impl From<CheckpointEvaluation> for Evaluation {
-    fn from(pair: CheckpointEvaluation) -> Self {
-        Evaluation {
-            loss: pair.loss,
-            accuracy: pair.accuracy,
-        }
-    }
-}
-
-impl From<&EvaluationSet> for CheckpointEvaluationSet {
-    fn from(eval: &EvaluationSet) -> Self {
-        CheckpointEvaluationSet {
-            train: eval.train.into(),
-            validation: eval.validation.map(Into::into),
-            test: eval.test.into(),
-        }
-    }
-}
-
-impl From<CheckpointEvaluationSet> for EvaluationSet {
-    fn from(evals: CheckpointEvaluationSet) -> Self {
-        EvaluationSet {
-            train: evals.train.into(),
-            validation: evals.validation.map(Into::into),
-            test: evals.test.into(),
-        }
-    }
-}
-
-/// A reference to a checkpoint subdirectory, named `checkpoint-{epoch:06}`.
-struct CheckpointRef {
-    epoch: usize,
-    dir: PathBuf,
-}
-
-/// Scans `dir` for `checkpoint-*` subdirectories, sorted by their numeric epoch
-/// (not lexically, so 10+ checkpoints sort correctly). Reads no other files.
-fn scan_checkpoints(dir: &Path) -> Result<Vec<CheckpointRef>> {
-    let mut checkpoints: Vec<CheckpointRef> = fs::read_dir(dir)?
-        .filter_map(StdResult::ok)
-        .filter_map(|entry| {
-            let path = entry.path();
-            let name = path.file_name()?.to_str()?;
-            if path.is_dir() {
-                let epoch = name.strip_prefix("checkpoint-")?.parse::<usize>().ok()?;
-                Some(CheckpointRef { epoch, dir: path })
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    checkpoints.sort_by_key(|s| s.epoch);
-    Ok(checkpoints)
 }
 
 /// A handle to a training run directory: its location and run-level metadata.
@@ -173,15 +83,13 @@ impl TrainingRun {
 
     /// Returns a recorder for writing checkpoints into this run.
     pub fn recorder(&self) -> CheckpointRecorder {
-        CheckpointRecorder {
-            dir: self.dir.clone(),
-        }
+        CheckpointRecorder::new(self.dir.clone())
     }
 
     /// Removes checkpoints whose epoch is greater than `from_epoch`, rewinding
     /// the trajectory. Returns the number of checkpoints removed.
     pub fn trim_after(&self, from_epoch: usize) -> Result<usize> {
-        let to_remove: Vec<CheckpointRef> = scan_checkpoints(&self.dir)?
+        let to_remove: Vec<_> = scan_checkpoints(&self.dir)?
             .into_iter()
             .filter(|checkpoint| checkpoint.epoch > from_epoch)
             .collect();
@@ -200,187 +108,12 @@ impl TrainingRun {
     }
 }
 
-/// Writes model checkpoints to a directory.
-///
-/// Each call to [`write`](CheckpointRecorder::write) creates a subdirectory
-/// `checkpoint-{epoch:06}/` containing `model.safetensors`, `evaluations.json`,
-/// and (if the optimizer or scheduler have internal state) `optimizer.safetensors`.
-/// Implements [`TrainingCallback`]: [`on_checkpoint`](TrainingCallback::on_checkpoint)
-/// writes a checkpoint. Obtained via [`TrainingRun::recorder`].
-#[derive(Debug)]
-pub struct CheckpointRecorder {
-    dir: PathBuf,
-}
-
-impl CheckpointRecorder {
-    /// Writes `checkpoint-{epoch:06}/` containing the model weights, evaluations,
-    /// and any optimizer/scheduler state.
-    pub fn write(
-        &self,
-        model: &NeuralNetwork,
-        optimizer: &dyn Optimizer,
-        scheduler: &dyn Scheduler,
-        evaluation: &EvaluationSet,
-        epoch: usize,
-    ) -> Result<()> {
-        let checkpoint_dir = self.dir.join(format!("checkpoint-{epoch:06}"));
-        fs::create_dir_all(&checkpoint_dir)?;
-
-        let mut entries = Vec::new();
-        let mut metadata = HashMap::new();
-        model.collect_tensors(&mut entries, &mut metadata);
-        tensors::save(checkpoint_dir.join("model"), entries, metadata)?;
-
-        json::save(
-            &CheckpointEvaluationSet::from(evaluation),
-            checkpoint_dir.join("evaluations"),
-        )?;
-
-        let mut state = optimizer.to_state().unwrap_or(OptimizerState {
-            tensors: Vec::new(),
-            metadata: HashMap::new(),
-        });
-        if let Some(scheduler_state) = scheduler.to_state() {
-            state.metadata.insert(
-                "scheduler.current_step".to_string(),
-                scheduler_state.current_step.to_string(),
-            );
-        }
-        if !state.tensors.is_empty() || !state.metadata.is_empty() {
-            optimizer_io::save(&state, checkpoint_dir.join("optimizer"))?;
-        }
-
-        Ok(())
-    }
-}
-
-impl TrainingCallback for CheckpointRecorder {
-    fn on_checkpoint(
-        &mut self,
-        model: &NeuralNetwork,
-        optimizer: &dyn Optimizer,
-        scheduler: &dyn Scheduler,
-        eval: &EvaluationSet,
-        epoch: usize,
-    ) -> Result<()> {
-        self.write(model, optimizer, scheduler, eval, epoch)
-    }
-}
-
-/// Lazy read access to a directory of checkpoints written by [`CheckpointRecorder`].
-///
-/// [`CheckpointArchive::load`] only scans directory names — no files are read until
-/// [`model_at`](CheckpointArchive::model_at) or
-/// [`evaluation_history`](CheckpointArchive::evaluation_history) is called.
-pub struct CheckpointArchive {
-    entries: Vec<CheckpointRef>,
-}
-
-impl CheckpointArchive {
-    /// Scans `dir` for `checkpoint-*` subdirectories, sorted by epoch.
-    pub fn load<P: AsRef<Path>>(dir: P) -> Result<Self> {
-        let dir = Path::combine_safe_with_cwd(dir)?;
-        Ok(CheckpointArchive {
-            entries: scan_checkpoints(&dir)?,
-        })
-    }
-
-    /// Returns the number of checkpoints found.
-    pub fn len(&self) -> usize {
-        self.entries.len()
-    }
-
-    /// Returns true if no checkpoints were found.
-    pub fn is_empty(&self) -> bool {
-        self.entries.is_empty()
-    }
-
-    /// Returns the absolute epoch number of the checkpoint at `index`, from the
-    /// scanned directory name (zero I/O).
-    pub fn epoch_at(&self, index: usize) -> Option<usize> {
-        self.entries.get(index).map(|s| s.epoch)
-    }
-
-    /// Loads the model at position `index` from its checkpoint directory.
-    pub fn model_at(&self, index: usize) -> Result<NeuralNetwork> {
-        let entry = self.entries.get(index).ok_or_else(|| {
-            Error::new(
-                InvalidData,
-                format!(
-                    "checkpoint index {index} out of range (archive has {} checkpoints)",
-                    self.entries.len()
-                ),
-            )
-        })?;
-
-        let bytes = secure_read(entry.dir.join("model.safetensors"))?;
-        let st =
-            SafeTensors::deserialize(&bytes).map_err(|e| Error::new(InvalidData, e.to_string()))?;
-        let metadata = tensors::read_metadata(&bytes)?;
-        NeuralNetwork::from_tensors(&st, &metadata)
-    }
-
-    /// Loads the optimizer/scheduler state at position `index`, or `Ok(None)`
-    /// if the checkpoint has no `optimizer.safetensors` (stateless optimizer,
-    /// or a checkpoint written before this state was tracked).
-    pub fn optimizer_at(&self, index: usize) -> Result<Option<OptimizerState>> {
-        let entry = self.entries.get(index).ok_or_else(|| {
-            Error::new(
-                InvalidData,
-                format!(
-                    "checkpoint index {index} out of range (archive has {} checkpoints)",
-                    self.entries.len()
-                ),
-            )
-        })?;
-
-        if !entry.dir.join("optimizer.safetensors").exists() {
-            return Ok(None);
-        }
-
-        optimizer_io::load(entry.dir.join("optimizer")).map(Some)
-    }
-
-    /// Reads all `evaluations.json` files into a pure [`EvaluationHistory`].
-    pub fn evaluation_history(&self) -> Result<EvaluationHistory> {
-        let mut history = Vec::with_capacity(self.entries.len());
-
-        for entry in &self.entries {
-            let evals: CheckpointEvaluationSet = json::load(entry.dir.join("evaluations"))?;
-            history.push(EpochEvaluation {
-                epoch: entry.epoch,
-                evaluation: evals.into(),
-            });
-        }
-
-        Ok(EvaluationHistory::new(history))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::activations::RELU;
-    use crate::model::NeuronLayerSpec;
-    use crate::optimizers::Adam;
-    use crate::schedulers::ConstantScheduler;
-    use crate::training::LearningRate;
-    use ndarray::Array2;
-
-    fn sample_optimizer() -> Adam {
-        Adam::with_defaults(LearningRate::new(0.01).unwrap())
-    }
-
-    fn sample_scheduler() -> ConstantScheduler {
-        ConstantScheduler::new(LearningRate::new(0.01).unwrap())
-    }
 
     fn temp_dir(tag: &str) -> PathBuf {
-        let dir = PathBuf::from(format!(
-            "target/nrn_checkpoint_{}_{}",
-            tag,
-            std::process::id()
-        ));
+        let dir = PathBuf::from(format!("target/nrn_run_{}_{}", tag, std::process::id()));
         fs::create_dir_all(&dir).unwrap();
         dir
     }
@@ -389,73 +122,41 @@ mod tests {
         let _ = fs::remove_dir_all(dir);
     }
 
-    fn create_run<P: AsRef<Path>>(
-        path: P,
-        dataset: &str,
-        overwrite: bool,
-    ) -> Result<CheckpointRecorder> {
-        Ok(TrainingRun::create(
-            path,
-            &TrainingMeta {
-                dataset: dataset.to_string(),
-                hyperparams: HyperParamsRecord::sample(),
-            },
-            overwrite,
-        )?
-        .recorder())
-    }
-
-    fn sample_model() -> NeuralNetwork {
-        let specs = NeuronLayerSpec::network_for(vec![3], &*RELU, 2);
-        NeuralNetwork::initialization(2, &specs)
-    }
-
-    fn make_eval(loss: f32, with_validation: bool) -> EvaluationSet {
-        EvaluationSet {
-            train: Evaluation {
-                loss,
-                accuracy: 0.5,
-            },
-            validation: with_validation.then_some(Evaluation {
-                loss: loss + 1.0,
-                accuracy: 0.1,
-            }),
-            test: Evaluation {
-                loss: loss + 100.0,
-                accuracy: 0.9,
-            },
+    fn meta(dataset: &str) -> TrainingMeta {
+        TrainingMeta {
+            dataset: dataset.to_string(),
+            hyperparams: HyperParamsRecord::sample(),
         }
+    }
+
+    /// Creates an empty `checkpoint-{epoch:06}/` directory. The run lifecycle
+    /// (create/trim) only cares about which checkpoint directories exist, not
+    /// their contents — that I/O is exercised in [`crate::io::checkpoint`].
+    fn make_checkpoint(dir: &Path, epoch: usize) {
+        fs::create_dir_all(dir.join(format!("checkpoint-{epoch:06}"))).unwrap();
     }
 
     #[test]
     fn meta_json_written_by_create() {
         let dir = temp_dir("meta");
-        create_run(&dir, "my_dataset", false).unwrap();
+        TrainingRun::create(&dir, &meta("my_dataset"), false).unwrap();
 
-        let meta = TrainingMeta::load(&dir).unwrap();
+        let loaded = TrainingMeta::load(&dir).unwrap();
         cleanup(&dir);
 
-        assert_eq!(meta.dataset, "my_dataset");
+        assert_eq!(loaded.dataset, "my_dataset");
     }
 
     #[test]
     fn create_errors_if_checkpoints_exist_without_overwrite() {
         let dir = temp_dir("no_overwrite");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
+        TrainingRun::create(&dir, &meta("ds"), false).unwrap();
+        make_checkpoint(&dir, 0);
 
-        let result = create_run(&dir, "ds", false);
+        let result = TrainingRun::create(&dir, &meta("ds"), false);
         cleanup(&dir);
 
-        let err = result.unwrap_err();
+        let err = result.err().unwrap();
         assert_eq!(err.kind(), AlreadyExists);
         assert!(err.to_string().contains("already exists"));
     }
@@ -463,29 +164,13 @@ mod tests {
     #[test]
     fn create_with_overwrite_purges_previous_checkpoints() {
         let dir = temp_dir("overwrite");
-        let recorder = create_run(&dir, "ds", false).unwrap();
+        TrainingRun::create(&dir, &meta("ds"), false).unwrap();
         for i in 0..3 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i * 10,
-                )
-                .unwrap();
+            make_checkpoint(&dir, i * 10);
         }
 
-        let recorder = create_run(&dir, "ds", true).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
+        TrainingRun::create(&dir, &meta("ds"), true).unwrap();
+        make_checkpoint(&dir, 0);
 
         let archive = CheckpointArchive::load(&dir).unwrap();
         cleanup(&dir);
@@ -494,144 +179,11 @@ mod tests {
     }
 
     #[test]
-    fn write_names_checkpoint_dir_by_epoch() {
-        let dir = temp_dir("write_epoch");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                37,
-            )
-            .unwrap();
-
-        let exists = dir.join("checkpoint-000037").is_dir();
-        cleanup(&dir);
-
-        assert!(exists);
-    }
-
-    #[test]
-    fn roundtrip_with_validation() {
-        let dir = temp_dir("roundtrip_val");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        for i in 0..3 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, true),
-                    i * 10,
-                )
-                .unwrap();
-        }
-
-        let history = CheckpointArchive::load(&dir)
-            .unwrap()
-            .evaluation_history()
-            .unwrap();
-        cleanup(&dir);
-
-        assert_eq!(history.len(), 3);
-        for (i, loss) in history.train_losses().iter().enumerate() {
-            assert_eq!(*loss, i as f32);
-        }
-        assert!(!history.validation_losses().is_empty());
-    }
-
-    #[test]
-    fn roundtrip_without_validation() {
-        let dir = temp_dir("roundtrip_noval");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        for i in 0..3 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i * 10,
-                )
-                .unwrap();
-        }
-
-        let history = CheckpointArchive::load(&dir)
-            .unwrap()
-            .evaluation_history()
-            .unwrap();
-        cleanup(&dir);
-
-        assert_eq!(history.len(), 3);
-        assert!(history.validation_losses().is_empty());
-    }
-
-    #[test]
-    fn predictions_survive_roundtrip() {
-        let dir = temp_dir("predictions");
-        let model = sample_model();
-        let inputs = Array2::from_shape_fn((2, 4), |(i, j)| (i + j) as f32 * 0.3);
-
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &model,
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let loaded = archive.model_at(0).unwrap();
-        cleanup(&dir);
-
-        assert_eq!(model.predict(inputs.view()), loaded.predict(inputs.view()));
-    }
-
-    #[test]
-    fn numeric_sort_beats_lexical() {
-        let dir = temp_dir("sort");
-        let model = sample_model();
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        for i in 0..12 {
-            recorder
-                .write(
-                    &model,
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i,
-                )
-                .unwrap();
-        }
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        cleanup(&dir);
-
-        assert_eq!(archive.len(), 12);
-        for i in 0..12 {
-            assert_eq!(archive.epoch_at(i), Some(i));
-        }
-    }
-
-    #[test]
-    fn resume_trims_checkpoints_after_from_epoch() {
+    fn trim_after_removes_checkpoints_past_from_epoch() {
         let dir = temp_dir("resume_trim");
-        let recorder = create_run(&dir, "ds", false).unwrap();
+        TrainingRun::create(&dir, &meta("ds"), false).unwrap();
         for i in 0..5 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i * 10,
-                )
-                .unwrap();
+            make_checkpoint(&dir, i * 10);
         }
 
         let trimmed = TrainingRun::open(&dir).unwrap().trim_after(20).unwrap();
@@ -644,19 +196,11 @@ mod tests {
     }
 
     #[test]
-    fn resume_keeps_all_when_from_epoch_is_last() {
+    fn trim_after_keeps_all_when_from_epoch_is_last() {
         let dir = temp_dir("resume_keep_all");
-        let recorder = create_run(&dir, "ds", false).unwrap();
+        TrainingRun::create(&dir, &meta("ds"), false).unwrap();
         for i in 0..3 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i * 10,
-                )
-                .unwrap();
+            make_checkpoint(&dir, i * 10);
         }
 
         let trimmed = TrainingRun::open(&dir).unwrap().trim_after(20).unwrap();
@@ -669,273 +213,26 @@ mod tests {
     }
 
     #[test]
-    fn resume_continues_writing_after_resume_point() {
+    fn trim_after_lets_a_checkpoint_be_repopulated() {
         let dir = temp_dir("resume_write");
-        let recorder = create_run(&dir, "ds", false).unwrap();
+        TrainingRun::create(&dir, &meta("ds"), false).unwrap();
         for i in 0..3 {
-            recorder
-                .write(
-                    &sample_model(),
-                    &sample_optimizer(),
-                    &sample_scheduler(),
-                    &make_eval(i as f32, false),
-                    i * 10,
-                )
-                .unwrap(); // 0, 10, 20
+            make_checkpoint(&dir, i * 10); // 0, 10, 20
         }
 
         let run = TrainingRun::open(&dir).unwrap();
-        run.trim_after(10).unwrap();
-        run.recorder()
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(99.0, false),
-                20,
-            )
-            .unwrap();
+        run.trim_after(10).unwrap(); // drops 20
+        make_checkpoint(&dir, 20); // rewritten
 
         let archive = CheckpointArchive::load(&dir).unwrap();
         cleanup(&dir);
 
-        assert_eq!(archive.len(), 3); // 0, 10, 20 (rewritten)
-    }
-
-    #[test]
-    fn empty_dir_archive_is_empty() {
-        let dir = temp_dir("empty");
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        cleanup(&dir);
-
-        assert!(archive.is_empty());
-        assert_eq!(archive.len(), 0);
-    }
-
-    #[test]
-    fn load_ignores_non_directory_checkpoint_file() {
-        let dir = temp_dir("non_dir_snap");
-        fs::write(dir.join("checkpoint-000000"), b"not a dir").unwrap();
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        cleanup(&dir);
-
-        assert!(archive.is_empty());
-    }
-
-    #[test]
-    fn load_ignores_checkpoint_dir_with_non_numeric_suffix() {
-        let dir = temp_dir("non_numeric_snap");
-        fs::create_dir_all(dir.join("checkpoint-abc")).unwrap();
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        cleanup(&dir);
-
-        assert!(archive.is_empty());
-    }
-
-    #[test]
-    fn model_at_out_of_range_gives_range_error() {
-        let dir = temp_dir("model_at_oob");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let msg = archive.model_at(99).unwrap_err().to_string();
-        cleanup(&dir);
-
-        assert!(msg.contains("out of range"), "got: {msg}");
-    }
-
-    #[test]
-    fn model_at_missing_model_file_fails() {
-        let dir = temp_dir("model_at_missing");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-        fs::remove_file(dir.join("checkpoint-000000").join("model.safetensors")).unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let result = archive.model_at(0);
-        cleanup(&dir);
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn model_at_corrupt_model_file_fails() {
-        let dir = temp_dir("model_at_corrupt");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-        fs::write(
-            dir.join("checkpoint-000000").join("model.safetensors"),
-            b"garbage",
-        )
-        .unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let result = archive.model_at(0);
-        cleanup(&dir);
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn evaluation_history_corrupted_evaluations_json_fails() {
-        let dir = temp_dir("corrupt_evals");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-        recorder
-            .write(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-        fs::write(
-            dir.join("checkpoint-000000").join("evaluations.json"),
-            b"not valid json",
-        )
-        .unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let result = archive.evaluation_history();
-        cleanup(&dir);
-
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn on_checkpoint_writes_a_checkpoint() {
-        let dir = temp_dir("on_checkpoint");
-        let mut recorder = create_run(&dir, "ds", false).unwrap();
-
-        recorder
-            .on_checkpoint(
-                &sample_model(),
-                &sample_optimizer(),
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                5,
-            )
-            .unwrap();
-
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        cleanup(&dir);
-
-        assert_eq!(archive.len(), 1);
-        assert_eq!(archive.epoch_at(0), Some(5));
-    }
-
-    #[test]
-    fn write_with_adam_writes_optimizer_state() {
-        let dir = temp_dir("optimizer_adam");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-
-        let mut optimizer = sample_optimizer();
-        let mut trained_model = sample_model();
-        let gradients = crate::gradients::Gradients {
-            dw: Array2::from_elem(trained_model.layers[0].weights.dim(), 0.1),
-            db: ndarray::Array1::from_elem(trained_model.layers[0].biases.dim(), 0.1),
-        };
-        optimizer.update(0, &mut trained_model.layers[0], &gradients);
-        optimizer.step();
-
-        recorder
-            .write(
-                &trained_model,
-                &optimizer,
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-
-        let exists = dir.join("checkpoint-000000/optimizer.safetensors").exists();
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let state = archive.optimizer_at(0).unwrap();
-        cleanup(&dir);
-
-        assert!(exists);
-        let state = state.unwrap();
-        assert!(!state.tensors.is_empty());
-        assert_eq!(
-            state.metadata.get("time_step").map(String::as_str),
-            Some("2")
-        );
-    }
-
-    #[test]
-    fn write_with_sgd_writes_no_optimizer_file() {
-        let dir = temp_dir("optimizer_sgd");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-
-        let optimizer =
-            crate::optimizers::StochasticGradientDescent::new(LearningRate::new(0.01).unwrap());
-        recorder
-            .write(
-                &sample_model(),
-                &optimizer,
-                &sample_scheduler(),
-                &make_eval(0.0, false),
-                0,
-            )
-            .unwrap();
-
-        let exists = dir.join("checkpoint-000000/optimizer.safetensors").exists();
-        let archive = CheckpointArchive::load(&dir).unwrap();
-        let state = archive.optimizer_at(0).unwrap();
-        cleanup(&dir);
-
-        assert!(!exists);
-        assert!(state.is_none());
-    }
-
-    #[test]
-    fn write_fails_when_evaluations_path_is_a_directory() {
-        let dir = temp_dir("write_evals_dir_conflict");
-        let recorder = create_run(&dir, "ds", false).unwrap();
-
-        // Pre-create "evaluations.json" as a directory so json::save's fs::write fails.
-        fs::create_dir_all(dir.join("checkpoint-000000").join("evaluations.json")).unwrap();
-
-        let result = recorder.write(
-            &sample_model(),
-            &sample_optimizer(),
-            &sample_scheduler(),
-            &make_eval(0.0, false),
-            0,
-        );
-        cleanup(&dir);
-
-        assert!(result.is_err());
+        assert_eq!(archive.len(), 3); // 0, 10, 20
     }
 
     #[test]
     fn create_rejects_path_traversal() {
-        let result = create_run("../../nrn_traversal_test", "x", false);
+        let result = TrainingRun::create("../../nrn_traversal_test", &meta("x"), false);
         assert!(result.is_err());
     }
 
@@ -951,12 +248,6 @@ mod tests {
         let result = TrainingRun::open(&dir);
         cleanup(&dir);
 
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn load_rejects_path_traversal() {
-        let result = CheckpointArchive::load("../../nrn_traversal_test");
         assert!(result.is_err());
     }
 }
