+++ core/tests/convergence.rs
@@ -31,7 +31,7 @@ fn xor_converges_to_low_loss() {
             None,
             CROSS_ENTROPY_LOSS.clone(),
             Box::new(Adam::with_defaults(LearningRate::new(0.1).unwrap())),
-            Box::new(ConstantScheduler::new(LearningRate::new(0.1).unwrap())),
+            Box::new(ConstantScheduler::from_value(0.1).unwrap()),
             GradientClipping::None,
             None,
             0.1,
@@ -73,7 +73,7 @@ fn xor_converges_with_mini_batch() {
             Some(2),
             CROSS_ENTROPY_LOSS.clone(),
             Box::new(Adam::with_defaults(LearningRate::new(0.01).unwrap())),
-            Box::new(ConstantScheduler::new(LearningRate::new(0.01).unwrap())),
+            Box::new(ConstantScheduler::from_value(0.01).unwrap()),
             GradientClipping::None,
             None,
             0.1,
@@ -119,7 +119,7 @@ fn three_class_converges_to_low_loss() {
             None,
             CROSS_ENTROPY_LOSS.clone(),
             Box::new(Adam::with_defaults(LearningRate::new(0.05).unwrap())),
-            Box::new(ConstantScheduler::new(LearningRate::new(0.05).unwrap())),
+            Box::new(ConstantScheduler::from_value(0.05).unwrap()),
             GradientClipping::None,
             None,
             0.1,
